refactor: create MailHistory when MailSentEvent is published
