--- src/main/kotlin/apply/application/MailHistoryService.kt
@@ -1,22 +1,0 @@
-package apply.application
-
-import apply.application.mail.MailData
-import apply.domain.mail.MailHistoryRepository
-import apply.domain.mail.getOrThrow
-import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
-
-@Transactional
-@Service
-class MailHistoryService(
-    private val mailHistoryRepository: MailHistoryRepository
-) {
-    fun findAll(): List<MailData> {
-        return mailHistoryRepository.findAll().map { MailData(it) }
-    }
-
-    fun getById(mailHistoryId: Long): MailData {
-        val mailHistory = mailHistoryRepository.getOrThrow(mailHistoryId)
-        return MailData(mailHistory)
-    }
-}
