+++ src/main/kotlin/apply/domain/mail/MailSentEvent.kt
@@ -1,0 +1,9 @@
+package apply.domain.mail
+
+import apply.application.mail.MailData
+
+data class MailSentEvent(
+    val mailData: MailData,
+    val succeedRecipients: List<String>,
+    val failedRecipients: List<String>
+)
